train_rguo/requirements.txt
albumentations==0.4.5
apex==0.1
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
efficientnet-pytorch==0.6.3
future==0.18.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imgaug==0.2.6
joblib==0.15.1
kiwisolver==1.2.0
llvmlite==0.32.1
matplotlib==3.2.1
networkx==2.4
nose==1.3.7
numba==0.49.1
numpy==1.18.4
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
openslide-python==1.1.1
pandas==1.0.3
Pillow==7.1.2
portalocker==1.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.0.0
tifffile==2020.5.11
timm==0.1.30
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tqdm==4.46.0
urllib3==1.25.9
